projeto3/requirements.txt
google-generativeai
Pillow
python-dotenv
streamlit